Move admin routes at front-end to merchant routes
